requirements.txt
urequests
beautifulsoup4
